Show Hotkeys in modal
